# specs/001-mcp2221a-filament-sensor/contracts/api-contract.yaml
openapi: 3.0.0
info:
  title: MCP2221A Filament Sensor Monitor API
  version: 0.1.0
  description: HTTP API for dual filament sensor monitoring system
servers:
  - url: http://localhost:5002
    description: Local monitoring server
paths:
...
  /config:
    get:
      summary: Get current configuration
      description: Returns sensor pin assignments and calibration settings
      operationId: getConfig
      responses:
        '200':
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'

    post:
      summary: Update configuration
      description: Modify sensor settings and calibration
      operationId: updateConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigurationUpdate'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /alerts:
    get:
      summary: Get unacknowledged alerts
      description: Returns list of active alerts requiring attention
      operationId: getAlerts
      parameters:
        - name: sensor_id
          in: query
          description: Filter by sensor ID
          schema:
            type: integer
            enum: [1, 2]
        - name: severity
          in: query
          description: Filter by severity level
          schema:
            type: string
            enum: [INFO, WARNING, ERROR, CRITICAL]
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
...
components:
  schemas:
...
    ConfigurationResponse:
      type: object
      required:
        - sensors
        - polling
      properties:
        sensors:
          type: array
          items:
            $ref: '#/components/schemas/SensorConfig'
        polling:
          $ref: '#/components/schemas/PollingConfig'
    SensorConfig:
      type: object
      required:
        - id
        - movement_pin
        - runout_pin
        - calibration_mm_per_pulse
        - enabled
      properties:
        id:
          type: integer
          enum: [1, 2]
        movement_pin:
          type: integer
          minimum: 0
          maximum: 3
        runout_pin:
          type: integer
          minimum: 0
          maximum: 3
        calibration_mm_per_pulse:
          type: number
          minimum: 0.1
          maximum: 10.0
          default: 2.88
        enabled:
          type: boolean
    PollingConfig:
      type: object
      required:
        - interval_ms
        - debounce_ms
      properties:
        interval_ms:
          type: integer
          minimum: 10
          maximum: 1000
          default: 100
        debounce_ms:
          type: integer
          minimum: 0
          maximum: 100
          default: 2
    ConfigurationUpdate:
      type: object
      properties:
        sensors:
          type: array
          items:
            $ref: '#/components/schemas/SensorConfig'
        polling:
          $ref: '#/components/schemas/PollingConfig'
    Alert:
      type: object
      required:
        - id
        - timestamp
        - sensor_id
        - event_type
        - severity
        - message
        - acknowledged
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        sensor_id:
          type: integer
          enum: [1, 2]
        event_type:
          type: string
          enum:
            - RUNOUT_DETECTED
            - FILAMENT_INSERTED
            - MOVEMENT_STARTED
            - MOVEMENT_STOPPED
            - SENSOR_DISCONNECTED
            - SENSOR_RECONNECTED
        severity:
          type: string
          enum: [INFO, WARNING, ERROR, CRITICAL]
        message:
          type: string
        acknowledged:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...